Move staticClient tests into directory.
